Separating crime version label creation logic
